Move Oid to own hpp/cpp files
